// static-project/wsc-pc-vis/client/.yalc/@youzan/ebiz-components/components/easy-form/types.ts
export interface IFieldStatus {
  disabled: boolean;
  visible: boolean;
}
// ...
type DatePickerType =
  | 'DatePicker'
  | 'QuarterPicker'
  | 'MonthPicker'
  | 'WeekPicker'
  | 'DateRangePicker'
  | 'TimePicker'
  | 'TimeRangePicker';
export type FormFieldsType =
  | 'Input'
  | 'NumberInput'
  | 'Select'
  | 'Checkbox'
  | 'Radio'
  | 'Switch'
  | 'ColorPicker'
  | DatePickerType
  | 'DateRangeQuickPicker'
  | 'Custom'
  | 'Plain'
  /** 特殊的内置类型 */
  | '__internal_group__'
  | '__internal_list__';
// ...
type EasyFormPlainType = {
  type: 'Plain';
  node: React.ReactNode;
} & IEasyFormBaseConfig;
export type EasyFormInternalListType = IForbidden_InternalListConfig & IEasyFormBaseConfig;
export type EasyFormInternalGroupType = IForbidden_InternalGroupConfig & IEasyFormBaseConfig;
export type EasyFormConfigValidType = (
  | IEasyFormInputConfig
  | IEasyFormNumberInputConfig
  | IEasyFormSelectConfig
  | IEasyFormCheckboxConfig
  | IEasyFormRadioConfig
  | IEasyFormCustomConfig
  | IEasyFormSwitchConfig
  | IEasyFormColorPickerConfig
  | IEasyFormDatePickerConfig
  | IEasyFormDateRangePickerConfig
) &
  IEasyFormBaseConfig;
export type EasyFormConfigType =
  | EasyFormPlainType
  | EasyFormConfigValidType
  | EasyFormInternalGroupType
  | EasyFormInternalListType;
// ...
export interface IEasyFormFieldRenderProps<Model, StatusModel, Config = EasyFormConfigType> {
  model: Model;
  statusModel: StatusModel;
  config: Config;
  addColon: boolean;
  disabled?: boolean;
  onChange(key: string, value: any): void;
}

interface IEasyFormFiledBaseProps {
  value: any;
  onChange(input: any): void;
}
export type EasyFormFieldProps<P> = IEasyFormFiledBaseProps & P;
/** 内部类型的数据格式，比如list和group类型，需要有自己的状态，需要额外定义成set类型 */
export type EasyFormInternalType<P extends BasicModel<any>> = FieldSetModel<{
  $$type: FieldModel<string>;
  status: FieldModel<Record<string, any>>;
  value: P;
}>;

// model types
export type NormalEasyFormModel = FieldModel<any>;
export type GroupEasyFormModel = FieldSetModel<Record<string, NormalEasyFormModel>>;
export type ListEasyFormModel = FieldArrayModel<any>;
export type EasyFormModelType = NormalEasyFormModel | GroupEasyFormModel | ListEasyFormModel;
// status type
export type NormalStatusModel = FieldModel<IFieldStatus>;